github-data-collect/requirements.txt
pandas==2.0.1
python-dotenv==1.0.0
PyGithub==2.1.1